The models use in python with django
With django virtural environment to link together the tree in linux os to use the python3 to manage the 
trees files and environment variables that are apart of url.py and work with setttings.py.  

This file shows the use of all the python files and what they are apart of in the django installation python and linux tree.